Add basic tests for BarChart and GanttChart
Tests include a check for validation that when a series is mapped to a
column name that the specified column actually exists. Generalizes the test
originally added in commit df5b8af and moves it from Contex.Plot to each
of the modules that implement Contex.PlotContent in order to provide
full coverage.