Allow Datasets to accept map data & store natively
When a list of maps is passed to Dataset.new/1, the maps are stored
natively instead of being sorted and converted to lists. Adds guards to
various column lookup functions in order to access columns by key.

In order to create a plot from a dataset with a map, a `series_mapping`
option must be passed to the constructor. This value must be a map with
keys for each of that plot type's required series.