Manage column/plot element mapping with Mapping
Provides a Mapping module and struct that manage the process of
associating dataset columns and plot elements. Each plot type is
responsible for specifying what its required and optional plot
elements are, then when a mapping is specified the provided
dataset columns are validated.

A Mapping also provides accessor functions which can be used to
fetch the value for a given plot element from a given dataset row.