Adding plot options to PointPlot and LinePlot (#57)
Adding support for the plot options, allowing PointPlot and LinePlot to show/hide the x/y axis.

I've added default plot options as requested. Note the default_plot_options are a copy of the original from the plot.ex file, except it's now a map instead of a keyword list (if i understand correctly, that's the goal and it's converted by Plot.parse_attributes/1 before the being sent to the underlying implementation).