Fix running the tests on later versions of elixir (#64)
At some point it seems that ex_unit stopped running tests where there
are multiple modules with the same name. In this case it appears that
the duplicate names were a copy-paste issue so I've fixed the names.

I've also fixed a few compilation warnings.

But it looks like the ContexLineChartTest needs to be finished or
adjusted since it's not really testing anything right now (except that
an error is not raised).